Reorg. game controller and update libstem_gamepad
