+++ internal/c/libqb/include/event.h
@@ -7,5 +7,6 @@ void error(int32_t error_number);
 void evnt(uint32_t linenumber, uint32_t inclinenumber = 0, const char *incfilename = NULL);
 
 extern uint32_t new_error;
+extern uint32_t qbevent;
 
 #endif
