--- internal/c/parts/input/game_controller/download/gamepad_1.4.0_pre1_source/version
@@ -1,3 +1,0 @@
-VERSION_MAJOR=1
-VERSION_MINOR=4
-VERSION_TWEAK=0
