+++ internal/c/parts/input/game_controller/libstem_gamepad/Gamepad_private.h
