--- internal/c/parts/input/game_controller/notes.txt
@@ -1,1 +1,0 @@
-http://forums.tigsource.com/index.php?topic=10675.80
