+++ licenses/license_libstem_gamepad.txt
@@ -1,3 +1,6 @@
+--------------------------------------------------------------------------------
+License of libstem Gamepad:
+
 Copyright (c) 2014 Alex Diener
 
 This software is provided 'as-is', without any express or implied
